TypeChecker: throw an error if a field cannot have its set synthesised.
For example, if a channel had one component M where M = A.{0} | B.{1} then
c?x?y is invalid, as it would require x to be bound to either A or B, which
requires a lookahead in the evaluator. Therefore, we prohibit this.

This is done by introducing a new constraint, Inputable which is true for
anything for which the input set could be synthesised. Most things are 
imputable, with the exception of yield types (i.e. partially constructed
datatypes).