Fix evaluation of prefixes w/ multiple dots in the last field
For instance, c?x.y.

Previously, the evaluator incorrectly bound the whole of the
value to x, rather than only the first pat. This is because
it detected that x.y was the last field and hence used
extensions for x, rather than oneFieldExtensions.

The solution is to also use the number of patterns in the
current field to help decide which extensions operator to
use.