Don't save processes bound to patterns in an environment.
In reality, computation of the top level of any process bound to a pattern
is extremely cheap as the ProcName will be returned almost immediately. Thus,
there is no benefit to saving them in the environment; in particular this
means that during compilation, we would keep the entire process subtree
around (once the thunk has been evaluated by the compiler).

Therefore, we 'cheat' slightly and rather than binding a pattern to its
value in the environment, we instead bind the pattern to a value that is
guaranteed to never be evaluated to a full tree of processes (by making it
monadic).

This sorts out the memory usage of the compiler, taking it from ~3GB to 
60MB on Bill's sudoku examples.