Remove the unsafePerformIO hack within the evaluator.
We used unsafePerformIO before in order to create new names during evaluation for patterns. In particular, if a pattern matched more than only name then we would create a name, bind it to the result of the whole expression, then create 'extractors' that would extract each component of the pattern. This ensured laziness was preserved, whilst also ensuring we only evaluate things once.

Rather than creating the name in the evaluator, we now do it during the desugar phase. This has a couple of small knock-on effects to certain types, but nothing major.