Allow sets of processes to be formed.
To avoid allowing process to be compared we introduce an extra type constraint Set that is implied by Eq. We then also change some of the types of the built in functions to only require Set, rather than Eq, but leave some as requiring Eq to avoid punching holes in the type system. See CSPM/Prelude.hs and CSPM/TypeChecker/Pat.hs for details of such constraints.

We only test for syntactic equality, not semantic equivalence.