Make sure processes and functions are always distinguished correctly.
We already ensured that processes and functions were distinguished based on their free variables in a number of cases, but there were a few bugs. Further, the implementation overlapped somewhat.

We now have a single 'ScopeIdentifier' that tracks the free variables that have been bound and allows disambiguation of both processes and functions. Further, we now correctly disambiguate nested functions and functions/processes that depend on the value of variables bound by generators (in statements etc).