Make extensions/productions have proper types.
This comes at a slight cost; extensions cannot be called on fully
constructed datatype or values as this would require introducing the
type of the empty dot list, which would make the type system
undecidable.

This requires a new type constraint CYieldable that holds whenever the
type is something that can be yielded by => (so only TEvent or
TDatatype).