Fix type checking of parameterless modules.
Previously, parameterless modules were treated as a single atomic block
for the purposes of type-checking, but not quite correctly. This alters
the way they are type checked so that parameterless modules are treated
simply as namespaces. This simplifies their usage.

The change is relatively simple: parameterless modules are declared as
binding only their module name, and depend on all declarations in the
module. Each of the declarations in the module is then included in the
standard list of declarations to type check.

This should fix several bugs with generalisation of types inside
modules.