Refactor the evaluator.
This is a large commit that refactors the evaluator so that it is able
to do some analysis before starting evaluation. This change along gives
a 10% performance improvement, but in addition, we are now able to
implement simple optimisations for CSPM evaluation.

The first such optimisation that this permits is an optimisation of
prefixes c?x -> P where P does not use any variable bound by the
prefix. We are now able to just compute the right-hand side once, rather
than once for each possible value.