Add a new time and space profiler.
Underneath this uses the Haskell profiler to profile the CSPm code. The
trick is simply to work out a way to assign Haskell call centres
dynamically so they can apply to CSPm identifiers instead. See the
comment in CSPM.Evaluator.Profiler for an explanation of how this works.

This also includes an executable that processes the resulting profiles.