Fix type checking of mutually recursive datatypes.
This could sometimes erroneously fail if the types were not comparable
for equality. This is because of the way we checked if the datatypes
were comparable for equality -- whenever type checking a single
definition, we would speculatively mark it as comparable for equality,
and then backtrack if necessary. However, this would not work for
mutually recursive datatypes, and we would always fail to identify them
as comparable for equality.

This commit modifies this to mark all mutually-recursive datatypes as
comparable for equality, and then backtracks en-masse if this fails.