Fix scripts that were erroneously passing the type checker.
This fixes a problem with patterns such as x.y <- e passing the type
checker where the type of e is less general than the inferred type of
x.y. For example, x.y was previously being inferred as of type a.b but
e could be of type Event.

This is fixed by treating such patterns in the same way as we treat
function calls.